# tools/kotohiro.openapi.yaml
openapi: 3.0.1
info:
  title: kotohiro
  description: ''
  version: 1.0.0
...
paths:
...
  /talksessions/{talkSessionID}:
    get:
      summary: トークセッションの詳細
      deprecated: false
      description: ''
      operationId: getTalkSessionDetail
      tags:
        - talk_session
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema: &ref_2
                $ref: '#/components/schemas/talkSession'
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: TalkSession
    put:
      summary: セッション編集
      deprecated: false
      description: ''
      operationId: editTalkSession
      tags:
        - talk_session
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                theme:
                  example: ''
                  type: string
                scheduledEndTime:
                  type: string
                  format: date-time
                  example: ''
                latitude:
                  type: number
                  description: 緯度
                  example: 0
                  nullable: true
                longitude:
                  type: number
                  description: 経度
                  example: 0
                prefecture:
                  type: string
                  description: 都道府県
                  example: ''
                  nullable: true
                city:
                  type: string
                  description: 市区町村
                  example: ''
                  nullable: true
                description:
                  type: string
                  description: 説明文
                  example: ''
                  nullable: true
                thumbnailURL:
                  description: サムネイルURL。文中から一番最初の画像URLをサムネとする。
                  example: ''
                  type: string
                restrictions:
                  type: array
                  items:
                    type: string
                  uniqueItems: true
                  description: 投稿制限のキー
                  example:
                    - demographics.city
              required:
                - theme
                - scheduledEndTime
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema: *ref_2
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
      security:
        - apikey-header-SessionId: []
      x-ogen-operation-group: TalkSession
...
  /talksessions/{talkSessionID}/analysis:
    get:
      summary: 分析結果一覧
      deprecated: false
      description: ''
      operationId: talkSessionAnalysis
      tags:
        - talk_session
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  myPosition: &ref_3
                    $ref: '#/components/schemas/userGroupPosition'
                  positions:
                    type: array
                    items: *ref_3
                  groupOpinions:
                    type: array
                    items:
                      type: object
                      properties:
                        groupName:
                          type: string
                        groupId:
                          type: integer
                        opinions:
                          type: array
                          items:
                            type: object
                            properties:
                              opinion: *ref_0
                              user: *ref_1
                              agreeCount:
                                type: integer
                              disagreeCount:
                                type: integer
                              passCount:
                                type: integer
                            required:
                              - opinion
                              - user
                              - agreeCount
                              - disagreeCount
                              - passCount
                      required:
                        - groupName
                        - opinions
                        - groupId
                required:
                  - positions
                  - groupOpinions
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: TalkSession
  /talksessions/{talkSessionID}/report:
    get:
      summary: セッションレポートを返す
      deprecated: false
      description: ''
      operationId: getTalkSessionReport
      tags:
        - talk_session
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: string
                required:
                  - report
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: TalkSession
...
  /auth/token/info:
    get:
      summary: JWTの内容を返してくれる
      deprecated: false
      description: ''
      operationId: oauth_token_info
      tags:
        - auth
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  aud:
                    type: string
                    description: Audience
                  exp:
                    type: string
                    description: 有効期限
                  iat:
                    type: string
                    description: 発行日時
                  iss:
                    type: string
                    description: 発行者
                  sub:
                    type: string
                    description: ユーザID
                  jti:
                    type: string
                    description: JWT ID
                  displayID:
                    type: string
                    description: ユーザーID
                  displayName:
                    type: string
                    description: ユーザー名
                  iconURL:
                    type: string
                    description: アイコンURL
                  isRegistered:
                    type: boolean
                    description: ユーザ登録済みか
                  isEmailVerified:
                    type: boolean
                required:
                  - aud
                  - exp
                  - iat
                  - iss
                  - sub
                  - jti
                  - isRegistered
                  - isEmailVerified
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              example:
                code: GENERAL-000
                message: INTERNAL SERVER ERROR.
          headers: {}
      security:
        - apikey-header-SessionId: []
      x-ogen-operation-group: Auth
...
components:
  schemas:
    talkSession:
      type: object
      properties:
        id:
          type: string
          description: トークセッションID
        theme:
          type: string
          description: テーマ
        description:
          type: string
          description: 説明
          nullable: true
        owner: *ref_1
        createdAt:
          type: string
          description: 作成日時
        scheduledEndTime:
          type: string
          description: 終了予定日時
        location: *ref_9
        city:
          type: string
          description: 市区町村
          nullable: true
        prefecture:
          type: string
          description: 都道府県
          nullable: true
        thumbnailURL:
          type: string
          description: サムネ画像
          nullable: true
        restrictions:
          type: array
          items: *ref_6
          description: セッションの参加制限
      required:
        - id
        - theme
        - owner
        - createdAt
        - scheduledEndTime
        - restrictions
...
    userGroupPosition:
      type: object
      properties:
        posX:
          type: number
        posY:
          type: number
        displayId:
          type: string
        displayName:
          type: string
        iconURL:
          type: string
          nullable: true
        groupName:
          type: string
        groupId:
          type: integer
        perimeterIndex:
          type: integer
          description: 境界ポイントのインデックス
      required:
        - posX
        - posY
        - displayId
        - groupName
        - displayName
        - groupId
...
  securitySchemes:
    apikey-header-SessionId:
      type: apiKey
      in: header
      name: SessionId
servers: []
security: []